pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "barings-oprisk-platform"
version = "0.1.0"
description = "End-to-end fraud and operational risk detection prototype inspired by the 1995 Barings Bank collapse."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Maykon Lincoln / Codex" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "streamlit>=1.38.0",
  "pandas>=2.2.0",
  "numpy>=2.0.0",
  "scikit-learn>=1.5.0",
  "joblib>=1.4.0",
  "pydantic>=2.8.0",
  "PyJWT>=2.9.0",
  "cryptography>=43.0.0",
  "pyyaml>=6.0.2",
  "matplotlib>=3.9.0",
  "httpx>=0.27.0",
  "pytest>=8.3.0",
]

[project.optional-dependencies]
explain = ["shap>=0.46.0"]
mlops = ["mlflow>=2.17.0"]
deep = ["tensorflow-cpu>=2.18.0"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

